requirements.txt
django
djangorestframework
scikit-learn
xgboost
joblib
pandas
numpy

